Fix multiple engine in Python editor bug (#11032) (#11033)
The source of the combo was set to a non-static property. This caused
a really strange behavior when multiple Python nodes were present in
the graph. In that case the engines would appear repeated multiple
times.

Changing the property to static fixed the issue and it actually makes
more sense.