Translate DS array to NET array for description (#11040)
This fixes a bug where the description of Curve.SplitByParameter in the
library would show up a description of an obsoleted overload.

The cause of the issue was the inability of XmlDocumentationExtensions
to handle conversion from primitve DS arrays to their equivalent .NET
arrays. This made the module unable to find the exact overload, so it
resorted to a fallback which returns the first overload available,
which in this case was the incorrect one.

The problem is fixed for this particular instance, as it is now able to
find the exact overload. Note, however, that the mechanism is not
bullet-proof, as there is no guaranteed way to translate back a type
from DS to the original .NET type.