DYN-3069-Coverage-DSOffice
I created two test methods in order to get more than 90% in the code coverage for the DSOffice folder. I needed to create two files for each test, the Dyn file is for testing the Data.ImportExcel under specific circunstances and the csv file is for executing the ImportCSV under specific circunstances also.

Right now the CACoverage job is discarding the tests in the DSOffice folder (a filter was used to discard this tests), then the coverage is very low. Then I will proceed to create another pull request for removing this filter in the CACoverage repo.