Fix ImportExcelFileShowArray.dyn file
When I created this file I used the Browse button from the File Path node to select a spreadsheet then when saving the dyn file the path was saved harcoded. Then when running a test in DynamoMSOfficeTests was failing due to the hardcoded path so I did a change and updated the dyn file in order to use a relative path.