Dyn 3069 code coverage DSOffice (2nd pull request) (#11091)
* DYN-3069-Coverage-DSOffice2

I added one test method that will execute the WorkSheet.ConvertToDimensionalArray  method, also I added an excel file in which some data will be written.

* Excel.WriteDataToExcelWorksheet test

I created a new test for the Excel.WriteDataToExcelWorksheet method. it is using the dyn file in order to write a null value in a excel file (just to reach a specific part in the WriteDataToExcelWorksheet method).

* Fix WriteToReadOnlyFile

Just for testing purposes I commented this line.