Parse long.MinValue in Design Script
Values of type long were parsed in a two-step process were the unsigned
value was parsed first and then multiplied by its sign if present. This
type of processing did not work for -9223372036854775808, as its
unsigned value is not a valid long. The issue is fixed by prepending
the '-' sign in case the number is negative before parsing it.