Style Workspace References consistently
Changed the styles of the workspace references extension window to be
more consistent with respect to Dynamo. Particularly, buttons that used
default WPF button style were switched to now use the same style as in
the "Geometry Scaling" window. Also, refresh is no longer a button, to
get rid of the undesired hover style, and the image foreground is
changed on hover.