Handle errors when loading application icon (#11314)
When loading the application icon to be displayed in the extension
windows, the method used would result in an unhandled error when the
application was started from a network folder. This error is now
handled, resulting only in the absence of the icon for the extension
window. Also, the method used to get the application path is changed
for a more reliable one.