Undock extension views that provide a window (#11309) (#11317)
Extension views that provided a window rather than a control, like
TuneUp, were not properly handled by docking/undocking. The content was
being cast to a type that was too specific so it got lost. Also,
opening the extension directly as a floating window would result in a
crash because the content was not being unwrapped from the window in
this case.