Make IntegerSlider64Bit the only AKA for migration
There were two classes that were declaring an 'AlsoKnownAs' attribute
for 'DSCoreNodesUI.Input.IntegerSlider' in 'CoreNodeModels.Input':
'IntegerSlider' and 'IntegerSlider64Bit'. Since the latter obsoletes
the former, it should be the only one declaring an 'AlsoKnownAs'
attribute. Note that currently both are trying to be registered, which
results in only one of them being registered but not really making any
guarantees of which one it is. This also generate an exception catched
internally but showed in the console.