Cherry picks for RC2.10.1 (#11436)
* Dock and Undock view extension with Dynamic extension ID as edge case (#11394)

* Initial Commit with settings null check

* Add unit test

* Add view extension loading warning

* Update message

* Initial Commit (#11366)

* Do not aggregate node execution analytics anymore (#11356)

* Do not aggregate node execution analytics anymore

* Update to use new event combo

* Update Comments

* Make IntegerSlider the only AKA for migration (#11360)

There were two classes that were declaring an 'AlsoKnownAs' attribute
for 'DSCoreNodesUI.Input.IntegerSlider' in 'CoreNodeModels.Input':
'IntegerSlider' and 'IntegerSlider64Bit'. While the latter obsoletes
the former, it is not possible to migrate to it from XML, so only 
'IntegerSlider' should continue declaring an 'AlsoKnownAs' attribute.
Note that currently both classes are trying to be registered, which
results in only one of them being registered but not really making any
guarantees of which one it is. This also generates an exception catched
internally but shown in the console.

* Revised icons for AnyTrue and AnyFalse nodes (#11361)

* Implement ToString for DS Dictionary (#11358)

This is done to provide meaningful output when a dictionary is provided
to nodes like 'String from Array' and 'String from Object'. These nodes
call built-in DS functions that end up calling ToString on the
dictionary, which until now used the default implementation inherited
from Object which just returned the class name.

Co-authored-by: Martin Misol Monzo <mmisol@users.noreply.github.com>
Co-authored-by: Laurence Elsdon <laurence.elsdon@matterlab.co>