Make Python console test more robust
The test assertions the test makes can fail. Sometimes the text 'Dynamo
is up to date.' can appear logged in the middle of the expected output.
The test is made more robust by relaxing the check using a regular
expression instead. This will allow random lines of text logged before,
after, and between the output of each node.