DYN-3513 - updating branch with latest code in master
When merging the master branch (after updating my forked repo), i got several conflics in the Resources file, then I had to re-add the resources.
In the DynamoView I added the necessary code to launch the Preferences window when the user click the option in the Dynamo menu.
Finally I did some minor changes in the Preferences view.