Dyn 3513 preferences modal dialog (#11534)
* Revert "First version with migration from packages.config to package reference (#11186)"

This reverts commit c7a5e493d3de3cf11005c7ec3525e75f58265657.

* Revert "Revert "First version with migration from packages.config to package reference (#11186)""

This reverts commit 5906c998575fa3548986a8ce62ab9adbc12b6754.

* Issues Type Predicter GitHub Action

I added a new workflow (issue_type_predicte.yaml) that is using the ML.NET model for predicting the issue type (source repo), so If a new issue is created in the Dynamo repo this workflow will run and will predict if is a Wishlist issue or not. If is a Wishlist issue it will be labeled as "Wishlist" and then another workflow will move the issue to the DynamoWishlist repo. If the issue is incomplete or is not valid the label "NotMLEvaluated" will be added to the issue.

Also I added two scripts more, one will return the issue body in a json string and the other one will clan the issue body removing sections not used like "Dynamo Version" or "Stack Trace"

* Fix for the GitHub token

When testing the issue predicter workflow the issues labeled as Wishlist were not moved to the DynamoWishlist repo due that the "Move Issue by labels" workflow failed. There was a problem with the PAT used to label the issue, I was using the wrong one (no triggers actions).

* DYN-3513 - new modal dialog for preferences

I added a new View(xaml) for the Preferences window also I added several entries in the Resources so they can be used in the PreferencesView.xaml. This changes only will satisfy the next requirements:
- The modal dialog should come with Preferences Title
- The modal dialog should come with Save Changes button
- The modal dialog should come with split lines
- Dynamo will be blocked until user exit that dialog

* DYN-3513 - new modal dialog for preferences

Updated the TabControl and TabItem because the Background property was set incorrectly in some places.

* DYN-3513 - updating branch with latest code in master

When merging the master branch (after updating my forked repo), i got several conflics in the Resources file, then I had to re-add the resources.
In the DynamoView I added the necessary code to launch the Preferences window when the user click the option in the Dynamo menu.
Finally I did some minor changes in the Preferences view.

* DYN-3513 - Fixing the drag & drop

In the title bar the drag & drop was not working correctly so I increased the width of the TextBlock.

* Removing changes in Assembly file

Removing changes in Assembly file

* DYN-3513 - Fixing code review comments

Based on the code review comments:
I moved the styles to the DynamoModern.xml
I added several comments in the PreferencesView.xaml

* DYN-3513 - Fixing code review comments

I removed one line in the csproj file because is not needed.