DYN-3518-Preferences-General Tab
I added several controls in the General Tab of the Preferences window, i had to add some new styles in the DynamoModern.xaml for the ComboBoxes and the RadioButtons because they are using very specific backgrounds and font colors.
Also I added 3 entries in resources so the ComboBoxes can be populated.