Dyn 3518 preferences general tab (#11542)
* Revert "First version with migration from packages.config to package reference (#11186)"

This reverts commit c7a5e493d3de3cf11005c7ec3525e75f58265657.

* Revert "Revert "First version with migration from packages.config to package reference (#11186)""

This reverts commit 5906c998575fa3548986a8ce62ab9adbc12b6754.

* Issues Type Predicter GitHub Action

I added a new workflow (issue_type_predicte.yaml) that is using the ML.NET model for predicting the issue type (source repo), so If a new issue is created in the Dynamo repo this workflow will run and will predict if is a Wishlist issue or not. If is a Wishlist issue it will be labeled as "Wishlist" and then another workflow will move the issue to the DynamoWishlist repo. If the issue is incomplete or is not valid the label "NotMLEvaluated" will be added to the issue.

Also I added two scripts more, one will return the issue body in a json string and the other one will clan the issue body removing sections not used like "Dynamo Version" or "Stack Trace"

* Fix for the GitHub token

When testing the issue predicter workflow the issues labeled as Wishlist were not moved to the DynamoWishlist repo due that the "Move Issue by labels" workflow failed. There was a problem with the PAT used to label the issue, I was using the wrong one (no triggers actions).

* DYN-3518-Preferences-General Tab

I added several controls in the General Tab of the Preferences window, i had to add some new styles in the DynamoModern.xaml for the ComboBoxes and the RadioButtons because they are using very specific backgrounds and font colors.
Also I added 3 entries in resources so the ComboBoxes can be populated.

* DYN-3518-GeneralTab-CodeReview

Based on Aaron comments:
- I removed the PreferencesWindowFontSizes and PreferencesWindowNumberFormats entries from the Resources files due that they already exists in the resources.
- Renamed some private properties.
- Change the Width value for the Number Format combox and remove the Height value. in the PreferencesView.xaml

* DYN-3518-GeneralTab-CodeReview

I removed hard-coded strings in the PreferencesView.xmal and were added to StaticResources in the resx file.
Also I added extra comments in the PreferencesView.xaml