DYN-3551 Moving the Save Changes button
Due that the previous PR related to creating the Preference panel was approved and merged a new Jira task (also a branch - PR) was created for moving the Save Changes button further to the right. Then I modified the margin of the Save Changes button to be aligned with the Close button at the top.