Dyn 3518 preferences window - color fixes (#11579)
* Revert "First version with migration from packages.config to package reference (#11186)"

This reverts commit c7a5e493d3de3cf11005c7ec3525e75f58265657.

* Revert "Revert "First version with migration from packages.config to package reference (#11186)""

This reverts commit 5906c998575fa3548986a8ce62ab9adbc12b6754.

* Issues Type Predicter GitHub Action

I added a new workflow (issue_type_predicte.yaml) that is using the ML.NET model for predicting the issue type (source repo), so If a new issue is created in the Dynamo repo this workflow will run and will predict if is a Wishlist issue or not. If is a Wishlist issue it will be labeled as "Wishlist" and then another workflow will move the issue to the DynamoWishlist repo. If the issue is incomplete or is not valid the label "NotMLEvaluated" will be added to the issue.

Also I added two scripts more, one will return the issue body in a json string and the other one will clan the issue body removing sections not used like "Dynamo Version" or "Stack Trace"

* Fix for the GitHub token

When testing the issue predicter workflow the issues labeled as Wishlist were not moved to the DynamoWishlist repo due that the "Move Issue by labels" workflow failed. There was a problem with the PAT used to label the issue, I was using the wrong one (no triggers actions).

* DYN-3518 Testing Changes

During the Quality Testing some issues were found in the General tab, so the next fixes were done:

- When a specific tab in the left tab is selected the blue color was updated to use a lighter one (#ADE4DE).
- The functionality of highlighting in blue the Tab when is selected was removed.
- When a tab is selected the Font weight is Bold otherwise will be Regular.