Add OpenXML nodes to import/export data
Two new nodes are added that allow to import and export data from an
Excel spreadsheet. These nodes uses the OpenXML SDK, so they do not
make use of the Excel application, thus can be used without Office
installed locally.

These nodes should behave pretty similarly to the existing Excel nodes.
In order to prove that, the tests for the existing nodes have been
copied and adapted for the new nodes.

At the function signature level, there are the following differences:

ImportOpenXML
- Provides inputs for 'startRow' and 'startColumn' (optional)

ExportOpenXML
- Inputs for 'startRow' and 'startColumn' are optional