Dyn 3541 preferences visual settings (#11636)
* Revert "First version with migration from packages.config to package reference (#11186)"

This reverts commit c7a5e493d3de3cf11005c7ec3525e75f58265657.

* Revert "Revert "First version with migration from packages.config to package reference (#11186)""

This reverts commit 5906c998575fa3548986a8ce62ab9adbc12b6754.

* Issues Type Predicter GitHub Action

I added a new workflow (issue_type_predicte.yaml) that is using the ML.NET model for predicting the issue type (source repo), so If a new issue is created in the Dynamo repo this workflow will run and will predict if is a Wishlist issue or not. If is a Wishlist issue it will be labeled as "Wishlist" and then another workflow will move the issue to the DynamoWishlist repo. If the issue is incomplete or is not valid the label "NotMLEvaluated" will be added to the issue.

Also I added two scripts more, one will return the issue body in a json string and the other one will clan the issue body removing sections not used like "Dynamo Version" or "Stack Trace"

* Fix for the GitHub token

When testing the issue predicter workflow the issues labeled as Wishlist were not moved to the DynamoWishlist repo due that the "Move Issue by labels" workflow failed. There was a problem with the PAT used to label the issue, I was using the wrong one (no triggers actions).

* DYN-3541-Preferences-VisualSettings

I added the UI Controls for the Visual Settings tab.
The next functionalities were implemented:
- Add new Styles
- Remove existing Styles
- Validate against existing Styles.
- Functionality for expand/collapse WPF Expanders

* DYN-3541-Preferences-VisualSettings-2

Moving some steps located in the PreferencesView constructor to a method.

* DYN-3541-Preferences-VisualSettings

There was an extra color in DynamoModern.xaml that is not used then I remove it.

* DYN-3541-Preferences-VisualSettings

I added a new Style for the buttons Save and Cancel, since they have a specific Style
Also I modified the Style for the Add Style button since it was not following the design in hig autodesk.

* DYN-3541-Preferences-VisualSettings

Also I added a Style for the color button picker, added one text resource so it can be used in the xaml and finally I indented the xaml code.

* DYN-3541-Preferences-VisualSettings Display Settings

I included the Display Settings sections and the two toggle buttons described in the AC
Also I did a modification so when the "Add Style" button is pressed the focus goes to the Group Name edit box.

* DYN-3541-Preferences-VisualSettings Code Review

- Removed the png image and instead use fontawesome
- Removed duplicated resources for Show Edges and  Isolate Selected Geometry
- Obsolete a enum in ChangeScaleFactorPrompt class
- Some changes in indentation and creating region for private properties

* DYN-3541-Preferences-VisualSettings Code Review2

Based on Aaron's comments I changed the visibility from public to internal in 4 methods located in PreferencesViewModel.