DYN-3578-Enable-VisualSetting
I removed some code that won't be used due that the ChangelScaleFactorPrompt window was deleted (now instead of the window we have 4 radio buttons in the Preferences -> Visual Settings -> Geometry Scaling section).
Also I deleted some events/commands used when the ScaleFactor changes due that the event won't be used.