Dyn 3578 enable visual settings (#11657)
* Revert "First version with migration from packages.config to package reference (#11186)"

This reverts commit c7a5e493d3de3cf11005c7ec3525e75f58265657.

* Revert "Revert "First version with migration from packages.config to package reference (#11186)""

This reverts commit 5906c998575fa3548986a8ce62ab9adbc12b6754.

* Issues Type Predicter GitHub Action

I added a new workflow (issue_type_predicte.yaml) that is using the ML.NET model for predicting the issue type (source repo), so If a new issue is created in the Dynamo repo this workflow will run and will predict if is a Wishlist issue or not. If is a Wishlist issue it will be labeled as "Wishlist" and then another workflow will move the issue to the DynamoWishlist repo. If the issue is incomplete or is not valid the label "NotMLEvaluated" will be added to the issue.

Also I added two scripts more, one will return the issue body in a json string and the other one will clan the issue body removing sections not used like "Dynamo Version" or "Stack Trace"

* Fix for the GitHub token

When testing the issue predicter workflow the issues labeled as Wishlist were not moved to the DynamoWishlist repo due that the "Move Issue by labels" workflow failed. There was a problem with the PAT used to label the issue, I was using the wrong one (no triggers actions).

* DYN-3578-Enable-VisualSetting

Enabling the Geometry Scaling section part 1

* DYN-3578-Enable-VisualSetting

I added the functionality for re-loading the ScaleValue in the Geometry Scaling section when the Preferences window is re-opened.
Also I set-up the binding functionality for the Render Precision section, so the precision will be reflected in the 3d Figure.

* DYN-3578-Enable-VisualSetting

I removed some code that won't be used due that the ChangelScaleFactorPrompt window was deleted (now instead of the window we have 4 radio buttons in the Preferences -> Visual Settings -> Geometry Scaling section).
Also I deleted some events/commands used when the ScaleFactor changes due that the event won't be used.

* DYN-3578-Enable-VisualSettings

This functionality was additional and I think is not needed since is not forcing to re-run the Graph for updating the Render Precision, so it's been deleted.

* DYN-3578-Enable-VisualSettings

Fixing issue that was producing when merging the updated master branch into my feature branch.

* DYN-3578-Enable-VisualSettings - Code Review

Based on Aaron comments instead of removing the event and the method I added the Obsolete decorator.