DYN-3606-Preferences-Expanders Functionality
I added the functionality in the Features tab of having just one Expander expander at one time (this functionality already exits in the Visual Settings expander).

Also I added the functionality of recording the current state of the Expander in a way that every time that the Preference window is closed and re-opened then the Expanders will remember the previous state. Due that this functionality was implemented by session then when the Dynamo app is closed  the previous Expander settings will be lost.