DYN-3606-Preferences-Expanders Code Review 3
Based on the code review comments:
I changed the PreferencesViewModel property from public to Internal.
I reorganized usings in the PreferencesViewModel.cs file.
In the PreferencesView.xaml.cs due that I found a way to find all the Expanders in the PreferencesView, I refactored some code in a way and now there is no need to describe each expander name for binding the IsExpanded property.