DYN-3606-Preferences-Expanders Code Review 3
I re-implemented the functionality of recording/loading the current expansion state of each expander in a List<ExpanderSettings>, now instead of recording each one only records the one expanded (only one Expander can be expanded per each tab), so the info for one is loaded when the Preferences window is opened and just one info is saved when the Expander expansion state is updated (thiis approach was suggested by Astul).
For implementing this functionality I added a Converter that will be binded in each Expander as Michael suggested.