DYN-3767 PopupWindow
The changes are basically the creation of the PopupWindow UI, also I included some classes that will be needed later for supporting the Step hierarchy in which the child classes are Survey, Tooltip, Welcome and ExitTour.

One of the functionalities included is that when you press the close button the popup is closed (as expected).

Three buttons were added in the PopupWindow but for Welcome screen just the "Start Tour" button is shown, the popup pointer is hidden. For Tooltips the navigation controls and the pointer are shown.
Also the CustomRichTextBox class was added to enable adding a TextBox in a Tooltip/Welcome screen with functionalities like showing hyperlinks, text in bold, bullet point and images.