DYN-3818-GuidedTours-MenuEntries
I added the "Get Started" submenu and the InteractiveGuide MenuItem, so we can show the Get Started popups.
Also in the DynamoView.xaml.cs I added the code needed for creating most of the popups for the GetStarted tour (the content of the popup is read from the Resources file, that's why I added several entries in the resources).