DYN-3768-SurveyWindowUI
I created 3 resources for the two titles of the SurveyWindow and one for the text content.
I modified the CustomRichTextBox because hyperlink names separated by spaces were not supported then I did some changes and now are working.
I added the RatingControl UI and the SurveyPopup UI and some used styles in the DynamoModern.xaml.
Finally I added a instance of the Survey in the DynamoView.xaml.cs and now a survey popup will be displayed just for testing purposes.