DYN-3834-GeoScaling-GraphRun (#11845)
* DYN-3834-GeoScaling-GraphRun

There was an issue that when the Geometry Scaling was updated in a very large graph the Preferences panel was blocked not allowing the user to close it. I think this behavior was because when the Preference panel was opened immediately a graph run was executed (without changing anything in the Preferences panel).
I was NOT able to reproduce the Preferences panel freezing behavior but we noticed that the graph execution was always happening when the Preferences panel was opened (when nothing was changed), then I did a fix in a way that the selected radio button (in geometry scaling section) is updated every time a new workspace is created or a dyn file is loaded (as is happening with the WorkspaceModel.ScaleFactor), so the graph run will be executed just when the ScaleFactor is changed.

* DYN-3834-GeoScaling-GraphRun Code Review

Unsubscribing from the WorkspaceSettingsChanged when the PreferencesViewModel is cleared.

* DYN-3834-GeoScaling-GraphRun Code Review

Making method private due that is just used in the same class.

* DYN-3834-GeoScaling-GraphRun Code Review

I did some changes so only when the ScaleFactor was changed a graph run will be executed when the Preferences dialog is closed

* DYN-3834-GeoScaling-GraphRun Code Review2

Replacing the call to ForceRunCancelCommand method for MarkNodesAsModifiedAndRequestRun due that the first one was executing the graph when the RunMode = Manual at the moment of updating the Geometry Scaling in the other hand with the MarkNodesAsModifiedAndRequestRun doesn't happen that behavior.