DYN-3847-Popup-Guided-Flow
I added a static class for events so that when the Next button or Back button are clicked we continue the steps flow.
I added a json file that is read every time that the "Get Started" tour is executed.
I added the classes Guide and GuideManager so it will read the guides from a json file and interact with the Guide and Steps.
I added the Name in the HelpMenu so we can position a Popup over the MenuItem.
I removed the functionality of creating the steps in the DynamoView.xaml.cs and also I added the handlers for Click event in the Back and Next button.
Finally I added a event in the Step class so every time a Step (popup) is closed this event will be raised.