DYN-3847-Popup-Guided-Flow Code Review
I modified the json file, specifically replaced the pointer_direction json entry by tooltip_pointer_direction, not sure if I should use camelcase or snake_case.
The ReadGuides function was renamed and split in several methods.
Also when creating a Survey we need to pass the DynamoViewModel in order to  set the IsRatingVisible property.