DYN-3817-ExitTourWindow-Popup
I created the view for the Exit Tour Window and also the functionality of showing the ExitTour ( this is shown only when one of the steps is closed so the tour is not finished - this doesn't apply for the Survey popup).
A resource string was added and is related to the ExitTour content so it can be localized later.