DYN-3817-ExitTourWindow-Popup Code Review
The ExitTourWindow was renamed to RealTimeInfoWindow.
I reorganized the RealTimeInfoWindow so doesn't depend of a Step (then it can be used in a more generic way).
I created the properties Width, Height and TextContent inside the RealTimeInfoWindow  so they can be set up from outside. The TextContent property contains the text that will be shown in the popup and it can be updated on runtime.