DYN-3908-OpenControls-ByStep
I modified the json file to include the UI Automation part
Also I created a new class called StepUIAutomation that will be used by a Step to fetch the information from the json file.
Finally I modified the GuidesManager class for including a method that will create the Automation information and the Guide class for UIAutomation action when the Next or Back button are pressed in the current Popup.