Fix For Loading Json File
Instead to have the json file path in the GuidesManager constructor I created two public properties that will contains the assembly path execution and the JSON files guides path, so we can verify in the unit testing that this paths exists and if the Guide will be created correctly.
Also I added a unit test that will validate that the json guides file exists and that the Guides can be created.