DYN-3908-OpenControls-ByStep
I modified the guides json file to include the UI Automation part
Also I created a new class called StepUIAutomation that will be used by a Step to fetch the information from the json file.
I modified the GuidesManager class for including a method that will create the Automation information and the Guide class for UIAutomation action that will be used when the Next or Back button are pressed in the current Popup.
Finally I added some event handlers for the RealTimeInfoWindow, (this popup will appear just when the current Tour is closed without finishing it). This event handlers will prevent to have a pile of RealTimeInfoWindow if several tours are not finished.