DYN-3918-Popup-Location
I added some method to update the Popups location for when the Dynamo window is resized or moved, also I fixed the same problem for the ExitTour popup.
In the DynamoModern,xml I removed the black border in popups due that this behavior was just temporary (previously we did't have a dark overlay).