DYN-4165-Conditional-Flows
First I added a draft of Step2, Step3 popups and also the resources for the text displayed in each popup.
I fix a bug in the CustumRichTextBox due that was not showing text in bold.
In Guide, GuidesValidationMethods and PreValidation classes I added the needed functionality for conditional flows, then if the Terms of Service has not been accepted yet then it will show a popup otherwise it will show a different popup.
In GuideManager class I added the functionality of reading the UIAutomation information as list.
Finally I modified the Step class for supporting the functionality of executing the PreValidation method.