DYN-4168-GuideBackground-Split
First I added two new classes, the first one CutOffArea will be used for storing information related to the section of the Overlay that will be removed so it will allow user interaction through it, the other one is HighlightArea that will be used for storing information related to the highlight rectangle that is shown over some elements like MenuItems, with this new classes we will be providing separated functionality for each one.
Some methods were renamed in the HostControlInfo, Guide and GuideManager classes and also some modifications were done in order to use the CutOffArea and HighlightArea classes.
The json file was updated so we can support the new classes added in the previous guides (Get Started guide) and also dividing the functionality (previously the cutoff area and highlight rectangle were implemented using the same info).
Due that some Windows/Controls are dynamic in the Packages tour we had to calculate the UIElement dynamically when showing the Popup, then in this case the Step class was modified to calculate the UIElement that will host the popup when the Menu is clicked also due that when clicking a menu a Popup is shown (so it was a different VisualTree that then DynamoView) I had to modify the MenuStyleDictionary adding a new Rectangle that will be using for highlighting sub menu items.
Finally I modified the GuideBackground for using the new classes created CutOffArea and HighlightArea (for this one I implemented a glow effect in the highlight rectangle that is also shown in the Figma design).