DYN-4337-Step3-PackageSearch
I fixed the popups from the Step3 with the right text format.
I modified some of the Methods for passing the Guides Flow because is needed when trying to remove/disable the automation done when calling the Show() method.
I noticed that some steps doesn't require the UIAutomation cleanup process due that the Opened window is reused for future Steps then I inserted a new flag in order to know if automation should be clean up or not.
I modified the process for launching the Package Search due that the TermsOfServiceView was modal and was stopping the execution, then now can be modal or not depending if a guide is being executed
I modified the Style of the caret in the Popup due that we are disabling the next button in a UIAutomation step but the color was like it was still enabled.