DYN-4337-Step3-PackageSearch
I added a resource for the Autodesk Sample Package.
I added the WindowName property in the HighlightRectArea due that in some cases we need to highlight controls that don't belong to the main window (DynamoView).
I modified the template adding a Grid for the Accept button in the TermsOfUseView Window due that we need to insert a Rectangle.