DYN-4142-GetStarted-Localization
The Guides were not starting when using a different language due that we were passing to the MainGuideManager.LaunchTour() method a resource string parameter that was translated previously, then the Guide was not found when searching by Guide.Name.
For fixing the issue we need to use a hardcoded string with the same value than the Guide.Name in the json file.