DYN-4337-Step3-PackageSearch Code Review
I've created ENUM for the actions and controltypes used in the UIAutomation then in this way we won't have hardcoded strings.
Also I added some lines for preventing the crash in the GetStarted guide due that the ExitGuideWindow is not used (so doesn't have memory).
Finally I changed some code due that was not compiling when updating my branch with the latest changes.